Split posting-list search context classes into their own files
The posting-list search context template classes all lived inside the
single postinglistsearchcontext.{h,hpp} pair, which had grown large and
mixed several unrelated concerns. Extract them into dedicated
header/impl files so each class is self-contained and easier to find,
include and maintain:

  - PostingListFoldedSearchContextT -> posting_list_folded_search_context.{h,hpp}
  - NumericPostingSearchContext     -> numeric_posting_search_context.{h,hpp,cpp}
  - StringPostingSearchContext      -> string_posting_search_context.{h,hpp,cpp}

The explicit template instantiations stay in postinglistsearchcontext.cpp;
the matching extern-template declarations move alongside the extracted
class. Callers (single/multi numeric/string post attributes and the
searchcontext test) are updated to include the new headers. This is a
pure code move with no behavioral change.
