+++ searchlib/src/vespa/searchlib/attribute/multistringpostattribute.hpp
@@ -5,6 +5,7 @@
 #include "multi_string_enum_search_context.h"
 #include "multistringpostattribute.h"
 #include "string_direct_posting_store_adapter.hpp"
+#include "string_posting_search_context.hpp"
 
 #include <vespa/searchcommon/attribute/config.h>
 #include <vespa/searchlib/query/query_term_simple.h>
