+++ searchlib/src/vespa/searchlib/attribute/numeric_posting_search_context.cpp
@@ -1,0 +1,5 @@
+// Copyright Vespa.ai. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
+
+#include "numeric_posting_search_context.h"
+
+#include "numeric_posting_search_context.hpp"
