+++ searchlib/src/vespa/searchlib/attribute/string_posting_search_context.cpp
@@ -1,0 +1,9 @@
+// Copyright Vespa.ai. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
+
+#include "string_posting_search_context.hpp"
+
+namespace search::attribute {
+
+// no code here yet
+
+} // namespace search::attribute
