+++ searchlib/src/vespa/searchlib/attribute/string_posting_search_context.h
@@ -1,0 +1,30 @@
+// Copyright Vespa.ai. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
+
+#pragma once
+
+#include "posting_list_folded_search_context.h"
+#include "postinglistsearchcontext.h"
+
+namespace search::attribute {
+
+template <typename BaseSC, typename AttrT, typename DataT>
+class StringPostingSearchContext
+    : public PostingSearchContext<BaseSC, PostingListFoldedSearchContextT<DataT>, AttrT> {
+private:
+    using ExecuteInfo = queryeval::ExecuteInfo;
+    using Parent = PostingSearchContext<BaseSC, PostingListFoldedSearchContextT<DataT>, AttrT>;
+    using RegexpUtil = vespalib::RegexpUtil;
+    using Parent::_enumStore;
+    // Note: Steps iterator one or more steps when not using dictionary entry
+    bool use_dictionary_entry(PostingListSearchContext::DictionaryConstIterator& it) const override;
+    // Note: Uses copy of dictionary iterator to avoid stepping original.
+    bool use_single_dictionary_entry(PostingListSearchContext::DictionaryConstIterator it) const {
+        return use_dictionary_entry(it);
+    }
+    bool use_posting_lists_when_non_strict(const ExecuteInfo& info) const override;
+
+public:
+    StringPostingSearchContext(BaseSC&& base_sc, bool useBitVector, const AttrT& toBeSearched);
+};
+
+} // namespace search::attribute
